Class and methods for agents
Agents accumulate assets based on a normal distribution with mean and variance set to 1. If agents form a coalition, their mean and variance become .75 and .25 -- so their mean assets decrease, but they become less volatile. If an agent's assets hit 0 or less than 0, they die. If an agent's assets fall below their threshold (which is initialized when an agent is created) then that agent can be "coerced", e.g., if another agent wants to join a coalition with them then they have to.