Create network class and methods
The network class creates a directed graph, populates it with agents from the Agent class, and creates a visualization of the graph where edges between nodes represent coalitions between agents. The methods in this class mainly change the representation of the nodes as agents interact, and tracks these changes through a list of adjacency matrices.